Refactor out pipeline rectpred tasks
Move all luigi Tasks for pipeline of rectpred into separate modules so
that we can access the underlying functions without having luigi
installed
